# app/docs/swagger.yaml
basePath: /
definitions:
  main.Data:
    properties:
      date:
        type: string
      id:
        type: integer
      sensorID:
        type: integer
      value:
        type: number
    type: object
  main.RoomModel:
    properties:
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      name:
        type: string
      sensors:
        items:
          $ref: '#/definitions/main.Sensor'
        type: array
      url:
        type: string
    type: object
  main.Sensor:
    properties:
      data:
        items:
          $ref: '#/definitions/main.Data'
        type: array
      description:
        type: string
      id:
        type: integer
      measurementUnit:
        type: string
      meshID:
        type: string
      name:
        type: string
      roomModelID:
        type: integer
    type: object
host: visense.f4.htw-berlin.de:8080
info:
  contact: {}
  description: This API provides information about 3D room models with associated
    sensors and their data.
  license: {}
  title: vi-sense BIM API
  version: 0.1.1
paths:
  /models:
    get:
      description: Query all avaiable room models.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.RoomModel'
            type: array
        "400":
          description: ok
          schema:
            type: string
        "500":
          description: ok
          schema:
            type: string
      summary: Query models
      tags:
      - models
  /models/{id}:
    get:
      description: Query a single room model by id with containing sensors.
      parameters:
      - description: RoomModelID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RoomModel'
        "400":
          description: ok
          schema:
            type: string
        "404":
          description: ok
          schema:
            type: string
        "500":
          description: ok
          schema:
            type: string
      summary: Query room model
      tags:
      - models
  /sensors:
    get:
      description: Query all avaiable sensors.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Sensor'
            type: array
        "400":
          description: ok
          schema:
            type: string
        "500":
          description: ok
          schema:
            type: string
      summary: Query sensors
      tags:
      - sensors
  /sensors/{id}:
    get:
      description: Query a single sensor by id with containing sensor data.
      parameters:
      - description: SensorId
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Sensor'
        "400":
          description: ok
          schema:
            type: string
        "404":
          description: ok
          schema:
            type: string
        "500":
          description: ok
          schema:
            type: string
      summary: Query sensor
      tags:
      - sensors
schemes:
- http
swagger: "2.0"